WIP: Make SparseArrays an extension
We shouldn't need to load all of the sparse arrays infrastructure unless the user is using sparse arrays. There are a few things that need to be done here:

1. KLU.jl code needs to be copied so we can default to KLU calls without requiring a user-level `using KLU`
2. Some functions are made which only get a useful dispatch after using SparseArrays is loaded, but that should be sufficiently safe

But one big issue is that Krylov.jl has a dependency on SparseArrays, so we'd either need to drop it as a dependency and a new Krylov library, or fix Krylov.jl to make it an extension.